+++ sp-server/src/main/java/net/minecraft/src/DispenserBehaviorFireworks.java
@@ -7,7 +7,7 @@ final class DispenserBehaviorFireworks extends BehaviorDefaultDispenseItem {
 	public ItemStack dispenseStack(IBlockSource par1IBlockSource, ItemStack par2ItemStack) {
 		EnumFacing var3 = BlockDispenser.getFacing(par1IBlockSource.getBlockMetadata());
 		double var4 = par1IBlockSource.getX() + (double) var3.getFrontOffsetX();
-		double var6 = (double) ((float) par1IBlockSource.getYInt() + 0.2F);
+		double var6 = (float) par1IBlockSource.getYInt() + 0.2F;
 		double var8 = par1IBlockSource.getZ() + (double) var3.getFrontOffsetZ();
 		EntityFireworkRocket var10 = new EntityFireworkRocket(par1IBlockSource.getWorld(), var4, var6, var8,
 				par2ItemStack);
