+++ sp-server/src/main/java/net/minecraft/src/EntityMagmaCube.java
@@ -94,7 +94,7 @@ public class EntityMagmaCube extends EntitySlime {
 	 * Causes this entity to do an upwards motion (jumping).
 	 */
 	protected void jump() {
-		this.motionY = (double) (0.42F + (float) this.getSlimeSize() * 0.1F);
+		this.motionY = 0.42F + (float) this.getSlimeSize() * 0.1F;
 		this.isAirBorne = true;
 	}
