+++ sp-server/src/main/java/net/minecraft/src/EntityMob.java
@@ -88,9 +88,9 @@ public abstract class EntityMob extends EntityCreature implements IMob {
 		if (var4) {
 			if (var3 > 0) {
 				par1Entity.addVelocity(
-						(double) (-MathHelper.sin(this.rotationYaw * (float) Math.PI / 180.0F) * (float) var3 * 0.5F),
+						-MathHelper.sin(this.rotationYaw * (float) Math.PI / 180.0F) * (float) var3 * 0.5F,
 						0.1D,
-						(double) (MathHelper.cos(this.rotationYaw * (float) Math.PI / 180.0F) * (float) var3 * 0.5F));
+						MathHelper.cos(this.rotationYaw * (float) Math.PI / 180.0F) * (float) var3 * 0.5F);
 				this.motionX *= 0.6D;
 				this.motionZ *= 0.6D;
 			}
