+++ sp-server/src/main/java/net/minecraft/src/ItemFirework.java
@@ -13,8 +13,8 @@ public class ItemFirework extends Item {
 	public boolean onItemUse(ItemStack par1ItemStack, EntityPlayer par2EntityPlayer, World par3World, int par4,
 			int par5, int par6, int par7, float par8, float par9, float par10) {
 		if (!par3World.isRemote) {
-			EntityFireworkRocket var11 = new EntityFireworkRocket(par3World, (double) ((float) par4 + par8),
-					(double) ((float) par5 + par9), (double) ((float) par6 + par10), par1ItemStack);
+			EntityFireworkRocket var11 = new EntityFireworkRocket(par3World, (float) par4 + par8,
+					(float) par5 + par9, (float) par6 + par10, par1ItemStack);
 			par3World.spawnEntityInWorld(var11);
 
 			if (!par2EntityPlayer.capabilities.isCreativeMode) {
