+++ sp-server/src/main/java/net/minecraft/src/ItemMinecart.java
@@ -23,8 +23,8 @@ public class ItemMinecart extends Item {
 
 		if (BlockRailBase.isRailBlock(var11)) {
 			if (!par3World.isRemote) {
-				EntityMinecart var12 = EntityMinecart.createMinecart(par3World, (double) ((float) par4 + 0.5F),
-						(double) ((float) par5 + 0.5F), (double) ((float) par6 + 0.5F), this.minecartType);
+				EntityMinecart var12 = EntityMinecart.createMinecart(par3World, (float) par4 + 0.5F,
+						(float) par5 + 0.5F, (float) par6 + 0.5F, this.minecartType);
 
 				if (par1ItemStack.hasDisplayName()) {
 					var12.func_96094_a(par1ItemStack.getDisplayName());
