+++ sp-server/src/main/java/net/minecraft/src/ItemSnow.java
@@ -28,8 +28,8 @@ public class ItemSnow extends ItemBlockWithMetadata {
 						&& par3World.checkNoEntityCollision(
 								var12.getCollisionBoundingBoxFromPool(par3World, par4, par5, par6))
 						&& par3World.setBlockMetadata(par4, par5, par6, var14 + 1 | var13 & -8, 2)) {
-					par3World.playSoundEffect((double) ((float) par4 + 0.5F), (double) ((float) par5 + 0.5F),
-							(double) ((float) par6 + 0.5F), var12.stepSound.getPlaceSound(),
+					par3World.playSoundEffect((float) par4 + 0.5F, (float) par5 + 0.5F,
+							(float) par6 + 0.5F, var12.stepSound.getPlaceSound(),
 							(var12.stepSound.getVolume() + 1.0F) / 2.0F, var12.stepSound.getPitch() * 0.8F);
 					--par1ItemStack.stackSize;
 					return true;
